Add backend user api login,signup,logout & save logintoken to cookies

+++ server.js
@@ -2,6 +2,7 @@ import express from 'express'
 import cookieParser from 'cookie-parser'
 import { bugService } from './services/bug.service.js'
 import { loggerService } from './services/logger.service.js'
+import { userService } from './services/user.service.js'
 const app = express()
 
 
@@ -9,8 +10,9 @@ app.use(express.static('public'))
 app.use(cookieParser())
 app.use(express.json())
 
-// LIST - get cars
+// LIST - get bug
 app.get('/api/bug', (req, res) => {
+
     const filterBy = {
         title: req.query.title,
         minSev: +req.query.minSev,
@@ -21,13 +23,16 @@ app.get('/api/bug', (req, res) => {
     bugService.query(filterBy)
         .then(bugs => res.send(bugs))
         .catch(err => {
-            loggerService.error('Cannot get cars', err)
-            res.status(500).send('Cannot get cars')
+            loggerService.error('Cannot get bugs', err)
+            res.status(500).send('Cannot get bugs')
         })
 })
 
-// ADD car
+// ADD bug
 app.post('/api/bug', (req, res) => {
+    const loggedinUser = userService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send('Cannot remove bug')
+
     console.log(req.body)
     const bugToSave = {
         title: req.body.title,
@@ -39,13 +44,16 @@ app.post('/api/bug', (req, res) => {
     bugService.save(bugToSave)
         .then(bug => res.send(bug))
         .catch((err) => {
-            loggerService.error('Cannot save car', err)
-            res.status(500).send('Cannot save car', err)
+            loggerService.error('Cannot save bug', err)
+            res.status(500).send('Cannot save bug', err)
         })
 })
 
-// UPDATE car
+// UPDATE bug
 app.put('/api/bug', (req, res) => {
+    const loggedinUser = userService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send('Cannot remove bug')
+
     const bugToSave = {
         _id: req.body._id,
         title: req.body.title,
@@ -56,18 +64,18 @@ app.put('/api/bug', (req, res) => {
     bugService.save(bugToSave)
         .then(bug => res.send(bug))
         .catch((err) => {
-            loggerService.error('Cannot save car', err)
-            res.status(500).send('Cannot save car', err)
+            loggerService.error('Cannot save bug', err)
+            res.status(500).send('Cannot save bug', err)
         })
 })
 
-// READ - get car
+// READ - get bug
 app.get('/api/bug/:bugId', (req, res) => {
     console.log('req.params:', req.params)
-     let visitedCount = req.cookies.visitedCount || 0
+    let visitedCount = req.cookies.visitedCount || 0
     visitedCount++
     res.cookie('visitedCount', visitedCount, { maxAge: 5 * 1000 })
-    if(visitedCount>3){
+    if (visitedCount > 3) {
         return res.status(401).send('Wait for a bit')
     }
     const { bugId } = req.params
@@ -79,8 +87,11 @@ app.get('/api/bug/:bugId', (req, res) => {
         })
 })
 
-// DELETE - remove car
+// DELETE - remove bug
 app.delete('/api/bug/:bugId', (req, res) => {
+    const loggedinUser = userService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send('Cannot remove car')
+
     // console.log('req.params:', req.params)
     const { bugId } = req.params
     bugService.remove(bugId)
@@ -93,6 +104,64 @@ app.delete('/api/bug/:bugId', (req, res) => {
 
 })
 
+// User API
+app.get('/api/user', (req, res) => {
+    userService.query()
+        .then(users => res.send(users))
+        .catch(err => {
+            loggerService.error('Cannot load users', err)
+            res.status(400).send('Cannot load users')
+        })
+})
 
-const port = 3333
+app.get('/api/user/:userId', (req, res) => {
+    const { userId } = req.params
+
+    userService.getById(userId)
+        .then(user => res.send(user))
+        .catch(err => {
+            loggerService.error('Cannot load user', err)
+            res.status(400).send('Cannot load user')
+        })
+})
+
+// Auth API
+app.post('/api/auth/login', (req, res) => {
+    const credentials = req.body
+
+    userService.checkLogin(credentials)
+        .then(user => {
+            if (user) {
+                const loginToken = userService.getLoginToken(user)
+                res.cookie('loginToken', loginToken)
+                res.send(user)
+            } else {
+                res.status(401).send('Invalid Credentials')
+            }
+        })
+})
+
+app.post('/api/auth/signup', (req, res) => {
+    const credentials = req.body
+    userService.save(credentials)
+        .then(user => {
+            if (user) {
+                const loginToken = userService.getLoginToken(user)
+                res.cookie('loginToken', loginToken)
+                res.send(user)
+            } else {
+                res.status(400).send('Cannot signup')
+            }
+        })
+})
+
+app.post('/api/auth/logout', (req, res) => {
+    res.clearCookie('loginToken')
+    res.send('logged-out!')
+})
+
+app.get('/**', (req, res) => {
+    res.sendFile(path.resolve('public/index.html'))
+})
+const port = process.env.PORT || 3030
 app.listen(port, () => loggerService.info((`Server listening on port http://127.0.0.1:${port}/`)))
